Aoristic Sum

An S4 class to represent an aoristic analysis results.

Slots

breaks

A RataDie vector giving the date break between time-blocks.

weights

A numeric vector.

groups

A character vector to store the group names (if any).

p

A numeric array giving the aorisitic probabilities.

Note

This class inherits from aion::TimeSeries: dates are internally stored as rata die.

Coerce

In the code snippets below, x is an AoristicSum object.

as.data.frame(x)

Coerces to a data.frame.

See also

Author

N. Frerebeau