Chronological Apportioning

Chronological Apportioning

apportion(object, ...)

# S4 method for matrix
apportion(
  object,
  s0,
  s1,
  t0,
  t1,
  from = min(s0),
  to = max(s1),
  step = 25,
  method = c("uniform", "truncated"),
  z = 2,
  progress = getOption("kairos.progress")
)

# S4 method for CountMatrix
apportion(
  object,
  t0,
  t1,
  from = NULL,
  to = NULL,
  step = 25,
  method = c("uniform", "truncated"),
  z = 2,
  progress = getOption("kairos.progress")
)

Arguments

object

An \(m \times p\) arkhe::CountMatrix object.

...

Currently not used.

s0

A length-\(m\) numeric vector giving the site beginning dates (in years CE).

s1

A length-\(m\) numeric vector giving the site end dates (in years CE).

t0

A length-\(p\) numeric vector giving the type beginning dates (in years CE).

t1

A length-\(p\) numeric vector giving the type end dates (in years CE).

from

A length-one numeric vector giving the beginning of the period of interest (in years CE).

to

A length-one numeric vector giving the end of the period of interest (in years CE).

step

A length-one integer vector giving the step size, i.e. the width of each time step for apportioning (in years CE; defaults to \(25\)).

method

A character string specifying the distribution to be used (type popularity curve). It must be one of "uniform" (uniform distribution) or "truncated" (truncated standard normal distribution). Any unambiguous substring can be given.

z

An integer value giving the lower and upper truncation points (defaults to \(2\)). Only used if method is "truncated".

progress

A logical scalar: should a progress bar be displayed?

References

Roberts, J. M., Mills, B. J., Clark, J. J., Haas, W. R., Huntley, D. L. & Trowbridge, M. A. (2012). A Method for Chronological Apportioning of Ceramic Assemblages. Journal of Archaeological Science, 39(5): 1513-20. doi: 10.1016/j.jas.2011.12.022 .

See also

Author

N. Frerebeau