kairos: Analysis of Chronological Patterns from Archaeological Count Data

A toolkit for absolute dating and analysis of chronological patterns. This package includes functions for chronological modeling and dating of archaeological assemblages from count data. It allows to compute time point estimates (e.g. Mean Ceramic Date) and density estimates of the occupation and duration of an archaeological site.

Details

Package:	kairos

Type:	Package

Version:	0.0.0.9000

License:	GPL-3

Package options

kairos uses the following options() to configure behavior:

kairos.progress: a logical scalar. Should progress bars be displayed?

See also

Author

Full list of authors and contributors (alphabetic order)

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

IRAMAT-CRP2A (UMR 5060)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France