Refine CA-based Seriation

Refine CA-based Seriation

Usage

seriate_refine(object, ...)

# S4 method for AveragePermutationOrder
seriate_refine(object, cutoff, margin = 1, axes = 1, n = 30, ...)

# S4 method for BootstrapCA
seriate_refine(object, cutoff, margin = 1, axes = 1, ...)

Arguments

object

A PermutationOrder object (typically returned by seriate_average()).

...

Further arguments to be passed to internal methods.

cutoff

A function that takes a numeric vector as argument and returns a single numeric value (see below).

margin

A length-one numeric vector giving the subscripts which the refinement will be applied over: 1 indicates rows, 2 indicates columns.

axes

An integer vector giving the subscripts of the CA axes to be used.

n

A non-negative integer giving the number of bootstrap replications.

Value

A RefinePermutationOrder object.

Details

seriate_refine() allows to identify samples that are subject to sampling error or samples that have underlying structural relationships and might be influencing the ordering along the CA space.

This relies on a partial bootstrap approach to CA-based seriation where each sample is replicated n times. The maximum dimension length of the convex hull around the sample point cloud allows to remove samples for a given cutoff value.

According to Peebles and Schachner (2012), "[this] point removal procedure [results in] a reduced dataset where the position of individuals within the CA are highly stable and which produces an ordering consistent with the assumptions of frequency seriation."

References

Peeples, M. A., & Schachner, G. (2012). Refining correspondence analysis-based ceramic seriation of regional data sets. Journal of Archaeological Science, 39(8), 2818-2827. doi:10.1016/j.jas.2012.04.040 .

See also

Author

N. Frerebeau