kairos: Analysis of Chronological Patterns from Archaeological Count Data

A toolkit for absolute and relative dating and analysis of chronological patterns. This package includes functions for chronological modeling and dating of archaeological assemblages from count data. It provides methods for matrix seriation. It also allows to compute time point estimates and density estimates of the occupation and duration of an archaeological site.

Details

Package:	kairos

Type:	Package

Version:	1.2.0

License:	GPL-3

Package options

kairos uses the following options() to configure behavior:

kairos.precision: an integer indicating the number of decimal places for years (defaults to \(0\)).

kairos.progress: a logical scalar. Should progress bars be displayed?

See also

Author

Full list of authors and contributors (alphabetic order)

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Ben Marwick	University of Washington, USA

Anne Philippe	Université de Nantes, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

Archéosciences Bordeaux (UMR 6034)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France