Event Plot

Produces an activity or tempo plot.

plot_event(object, ...)

# S4 method for DateEvent
plot_event(
  object,
  type = c("activity", "tempo"),
  event = FALSE,
  select = 1,
  n = 500
)

Arguments

object
An object of class DateEvent to be plotted.

...
Currently not used.

type
A character string indicating the type of plot. It must be one of "activity" (default) or "tempo". Any unambiguous substring can be given.

event
A logical scalar: should the distribution of the event date be displayed? Only used if type is "activity".

select
A numeric or character vector giving the selection of the assemblage that are drawn.

n
A length-one non-negative numeric vector giving the desired length of the vector of quantiles for density computation.

Value

A ggplot2::ggplot object.

Event and Acccumulation Dates

plot_event() plots the probability estimate density curves of archaeological assemblage dates (event and accumulation dates; Bellanger and Husi 2012). The event date is plotted as a line, while the accumulation date is shown as a grey filled area.

The accumulation date can be displayed as a tempo plot (Dye 2016) or an activity plot (Philippe and Vibet 2017):

Tempo plot

A tempo plot estimates the cumulative occurrence of archaeological events, such as the slope of the plot directly reflects the pace of change.

Activity plot

An activity plot displays the first derivative of the tempo plot.

References

Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and Interpreting Archaeological Contexts Using Pottery. Journal of Archaeological Science, 39(4), 777-790. doi: 10.1016/j.jas.2011.06.031 .

Dye, T. S. (2016). Long-Term Rhythms in the Development of Hawaiian Social Stratification. Journal of Archaeological Science, 71, 1-9. doi: 10.1016/j.jas.2016.05.006 .

Philippe, A. & Vibet, M.-A. (2017). Analysis of Archaeological Phases Using the R Package ArchaeoPhases. Journal of Statistical Software, Code Snippets, 93(1), 1-25. doi: 10.18637/jss.v093.c01 .

See also

Author

N. Frerebeau

Examples

## Event and accumulation dates (Bellanger et al.)
## See the vignette:
if (FALSE) {
utils::vignette("kairos")
}